IPC-316: Put L2+ support behind a feature flag and remove release() cross-net fee (#319)
Fixes #316, #317.

This is the first PR for the "Hardening Cross-net Messages"
implementation for M3.
* It includes a new `l2PlusSupport` that limits the deployment of L2+
subnets.
* Removes the minimum fee required for release.
* Disables general-purpose cross-net message operations.

## Question for reviewers.
* I've commented some lines in the tests because they may be useful in
the future we decide to re-enable these features (should I remove
them?).
* We still have
`[minCrosNetFee](https://github.com/consensus-shipyard/ipc-solidity-actors/blob/9c85f9d110ba10737d258094085e70bb9f0ad54d/src/GatewayDiamond.sol#L23)`
as a construct parameters for the gateway. We currently not use this
value anymore. Should we keep it for future reference or directly remove
it (as it may be confusing).

## Next

In the next PR: 
* I will decouple checkpoints from cross-net execution. 
* I will disable relayer rewards to adapt to the current design where
there are no fees or incentives for them (at least for now)

---------

Signed-off-by: Alfonso de la Rocha <adlrocha@tutamail.com>
Co-authored-by: auto-commit <auto-commit@users.noreply.github.com>