IPC-318: Decouple checkpoint from messages (#326)
Builds upon: #319 
Part 1 of impl. for
https://github.com/consensus-shipyard/ipc-solidity-actors/issues/318

This PR is the second part of "Hardening Cross-net messages", it was
getting too long so I decided to split it into several PRs. This one:
- Decouples messages from checkpoint submission.
- Abstract the logic to submit signatures and create quorum over objects
so it can be re-used by bottom-up message batches in the next PR.
- Fixes a bug where epochs and height were being typed as `uint64` when
in Solidity it is a `uint256` (I feel this was dragged from the
translation of the Rust actors into Solidity, as `ChainEpoch` in
Filecoin is represented through a `uint64`).


## Next
- Implement the creation and submission of `QuorumBundles` for the
execution of bottom-up messages.

Note that this change may include several breaking changes to the
deployment scripts, Fendermint and IPC (but we should probably address
those once the monorepo is ready to make our lives easier).

---------

Signed-off-by: Alfonso de la Rocha <adlrocha@tutamail.com>
Co-authored-by: auto-commit <auto-commit@users.noreply.github.com>
Co-authored-by: Denis Kolegov <dnkolegov@gmail.com>