ERC20: enable a subnet's supply to be determined by an ERC20 token in the parent. (#313)
## Context

Various IPC users seek to create dedicated crypoeconomies backed by
their own token. They'd like to launch their token on another network
and inject it into their subnet as native circulating supply, so it can
be used (a) to pay for gas, and (b) as the native coin for value
transfers.

## Contribution

This PR introduces the notion of the "supply source" at the parent
subnet. This configuration object determines the kind of supply backing
a child subnet. It can be "native" or "ERC20". In the latter case, the
token address must be provided. We verify that the address indeed
corresponds to an ERC20 token.

We also add a `funcWithToken()` external non-payable function to the
`GatewayManagerFacet`. This is analogous to `fund()`, but for token
supply (where the amount must be passed as a parameter).

We block message propagation (when the feature flag is enabled) whenever
ERC20 networks are involved, as a sender, receiver, or intermediary
network.

## Non-goals

This PR does NOT introduce the ability to use an ERC20 token as
validator stake. We don't expect that feature to be useful at all, since
a major benefit of IPC is that subnets draw economic security from the
value of the collateral locked, which is denominated in the root coin.

## Design doc

https://www.notion.so/pl-strflt/ERC20-Native-coin-conversion-eab9267445b6411d96ec17ef02a6510e?pvs=4

## TODO

- [x] Prevent horizontal subnet xnet when subnet is using ERC20 supply.
- [x] Releasing/withdrawing funds.
- [x] ~Unit tests.~ (preferring integration tests with a real ERC20
token involved)
- [x] Integration tests (with TODOs for pending test cases)