fix(topdown): pull effects up until committed finality.
Fendermint was pulling parent effects only until committed_finality-1.

This choice assumed that chains like Filecoin leaked deferred execution
mechanics through the query mechanism we use (Eth JSON-RPC API). Therefore
it was deemed that the safe choice was to lag behind by one epoch. However,
we can safely assume that Eth JSON-RPC APIs will only expose blocks that
have been executed and whose side-effects are available (unless "pending"
is requested). Filecoin reconciles its deferred execution model with
Eth client expectations of immediate execution, alas with some known rough
edges [1].

Furthermore, this choice assumed that our chain_head_delay can be 0, i.e.
we can consume the exposed HEAD of the parent in real time. In practice,
that's not a good idea, ever.

All in all, this is problematic because it's a hidden. Users inspecting
the committed parent finality and seeing N now need to understand that
events emitted in N are (counterintuitively) not imported yet, and they
need to wait longer.

This PR:
- Rectifies the queried range so that we import up to, and including,
  the committed parent finality.
- Disallows a zero chain_head_delay.

Note: I'm clarifying with @aakoshh the mechanics of Fendermint's Eth
JSON-RPC API; concretely, if it renconciles deferred execution or not.
There are some off-by-one shifts assumed in a number of places, so
unclear. However, by disallowing a chain_head_delay of 0, we are able
to work with APIs that expose heights before their results are available.

Note 2: when we adopt light clients to watch the parent, some assumptions
may change.

[1] https://github.com/filecoin-project/lotus/issues/11205