fix(ci): contracts-prettier: use github.ref var.
github.head_ref refers to the name of the source branch of the
pull request. When the PR is opened from a fork, it refers to the
original branch ref _in the forked repo_. Instead, we want to
checkout the merge branch in our repo. github.ref is the right
variable to use for that, and works with local and external
pull requests.