chore: remove Python requirement for contracts development.
In the past, we didn't have a clean and structured Hardhat usage
for contract development/deployment, so we resorted to an adhoc
Python script to generate a selector library utility for testing
diamonds.

Our usage of Hardhat has improved significantly, so it's time to
integrate this codegen step as a Hardhat task in order to reduce
codebase entropy.