feat: Re-enable direct uploads in objects API (#453)
Enables direct, multi-part uploads in the objects API.

Currently, the only way to stage an object in a validator's Iroh node
is using reverse upload.
In this model, the client spawns a new IROH node add the blob into it.
Then it sends the address of its node to the validator.

There are two problems with this approach:
- It doesn't work in the browser because IROH doesn't run in the browser
  yet.
- It add the burden on the client to keep its IROH node up and running
  until the validator has downloaded the data.

The direct upload will allow the client to directly upload the data
using a POST request.

This PR also removes signature verification because it is not used
anymore.
Earlier we used to the signature verification to check the signer's
balance. We don't do this here anymore.

Closes #439

Signed-off-by: avichalp <hi@avichalp.me>