fix: convert blob.subscription into Hamt (#520)
## Overview
The `Blob` stuct was an ipld hamt that contained a HashMap of its
`Subcribers`, which contained a HashMap of `SubscriptionGroups`. These
nested hash maps had the potential to grow larger than can be flushed to
the blockstore (1MB).
This change converts the nested hash maps into nested hamts, hence
preventing issues with exceeding the 1MB limit

## Details 
This change include a lot of conversion of hash map to hamt while
maintaining logic. The majority of this is working through when to flush
the inner hamts and converting from `Option` to `Result`. It also
includes some additions to the hamt implementation. A hamt now has an
`iter` method so it can be used in for loops, and the hamt can now be
keyed with a `String`, which is what the subscription groups are keyed
with.

fixes #441

---------

Signed-off-by: Sander Pick <sanderpick@gmail.com>
Co-authored-by: Sander Pick <sanderpick@gmail.com>