feat: bundle and tag all related blobs (#553)
Resolves https://github.com/recallnet/entanglement/issues/25

After uploading a blob and entangling it a new hash sequence is being
created that bundles all relevant blob hashes: for original blob,
metadata blob and parity blobs.
Instead of returning an original blob's hash we return the hash of the
hash sequence.

During the upload process all the blobs are assigned auto tags that are
deleted in the end of the upload. The hash sequence is being assigned a
"temp-{hash_seq_hash}" tag that will eventually be replaced with
"stored-{hash_seq_hash}" by the validator.

---------

Co-authored-by: Friedel Ziegelmayer <me@dignifiedquire.com>