fix: bucket list for_each_ranged (#584)
# Summary

Closes #564 

- Changes the bucket state to use the `hamt` from `recall_ipld`
- Implements a custom `for_each_ranged` that allows skipping elements
that don't match a prefix

# Details

The easiest way to understand the behavior is to upload two objects
(`hello` and `test/hello`). I've done that in `testnet`. You can try.
Querying `testnet`,

```
recall bu query -a 0xff000000000000000000000000000000000000fd --prefix test --limit 1
{
  "objects": [],
  "common_prefixes": [],
  "next_key": "test/hello"
}
```

I queried by prefix with limit 1, and it returned nothing, although we
know there's something there. We can only get that element by doing
another call.

With this PR, we'll return the correct element in just one call because
we skip the element (`hello`) that does not match the prefix.

There's still one thing that could be strange

```
recall bu query -a 0xff000000000000000000000000000000000000fd --limit 1              
{
  "objects": [
    {
      "key": "hello",
      "value": {
        "hash": "z5oyjijptdwp3jnxqi3m7nmepzzdjqc5yxpperxfk4xnohnwaguq",
        "size": 6,
        "expiry": 1372597,
        "metadata": {
          "content-type": "text/plain"
        }
      }
    }
  ],
  "common_prefixes": [],
  "next_key": "test/hello"
}
```

I queried without prefix. It returned the correct element, but it also
returned the `next_key`. It should have been `null`, because we know
there's nothing there. That's because "test/hello" comes next in the
`hamt`. TBH I don't know if there's an easy solution to this.

In short, this whole issue arises because we're not really doing an
indexed search on the prefix, we're querying the hamt and then seeing if
the prefix matches or not.

cc @asutula

---------

Signed-off-by: Bruno Calza <brunoangelicalza@gmail.com>