refactor: improve resolve pool logic (#591)
- Attempts to address https://github.com/recallnet/ipc/issues/588 by,
- Instead of just logging an error, actually vote "nay" if there is some
issue with the resolve pool channel. This should at least un-stick
pending blobs that are not longer in the pool. I still don't know why
the channel would close, but I added some more logging to help figure
this out.
- Hydrate pending blobs from chain state if the local resolve pool is
empty. This allows a validator that restarts to vote on currently
pending blobs.
- Ensure that the local resolve pool does not exceed the pool's
configured concurrency. Previously, we were using concurrency to
determine how many new `added` blobs to move to `pending` without
considering the current size of the pool
- Reverts the addition of `IrohManager` to the Object API that was
causing the "Iroh node is not running" errors.

---------

Signed-off-by: Sander Pick <sanderpick@gmail.com>