fix: make fendermint blobs types public (#595)
The changes in #582 moved the `AccountInfo` type from the
`fendermint_blobs_shared` packages to the `fendermint_blobs` package.
This type is needed by the rust SDK, and the move made the type private.
The change in this PR makes the `fendermint_blobs::state` module public,
so the SDK can access it.
Related PR for the SDK is here:
https://github.com/recallnet/rust-recall/pull/256