feat: upgrade to iroh@0.34 (#565)
Upgrade iroh to the latest version of `0.34`.

See #557 for the original design discussion.

Closes #557 

## Current Design

https://number-zero.notion.site/recall-iroh-0-33-upgrade-1b95df1306fb8066be57dfa7b49fae1f?pvs=4

## Notes

This includes a breaking wire change for iroh, so all participants need
to use the latest version of this, once merged.
The change is the usage of raw public keys in the TLS setup, for details
you can check out
https://www.iroh.computer/blog/iroh-0-34-0-raw-public-keys

## Security Concerns

The RPC api exposed from the resolver, must not be exposed to the
public, as it has no authentication. So care needs to be taken to
restrict access accordingly.

## Related PRs

- [x] SDK: https://github.com/recallnet/rust-recall/pull/232
- [x] Entanglement on https://github.com/recallnet/entanglement/pull/37

## Open Tasks

- [x] green CI
- [x] cleanup
- [x] fix syscall interface
- [x] check for places that the new blobs tags API can be used

---------

Signed-off-by: Sander Pick <sanderpick@gmail.com>
Co-authored-by: Sander Pick <sanderpick@gmail.com>