refactor: blobs actor layout (#575)
The blobs actor state is a bit of a nightmare--not DRY, huge files,
giant logic blocks. This makes it error prone and hard to reason about
it. This is mostly because we constrained ourselves to dealing with
mutable references from hash maps. Now that everything is
HAMTified, we can clean this up.

While working on #443, I couldn't resist starting the cleanup.

- Adds an actor `Caller` type for managing the incoming message origin /
caller / from / sponsor (make all this DRY)
- Adds a state `Caller` for managing accounts, approvals, and account
operations like debiting credits (make all this DRY)
- Breaks up the giant state file into reasonable chunks: blobs, credit,
accounts. Each has a file for state, params, and methods.
- Parameterize all methods with too many args.
- Closes https://github.com/recallnet/ipc/issues/446
- Closes https://github.com/recallnet/ipc/issues/445
- Closes https://github.com/recallnet/ipc/issues/452

---------

Signed-off-by: Sander Pick <sanderpick@gmail.com>